Move SyncAll trait and implementations to strat_engine/device.rs
We are using SyncAll not just in backstore/metadata.rs, so hoist it up
a bit.
