+++ src/engine/strat_engine/backstore/mod.rs
@@ -14,5 +14,5 @@ mod setup;
 mod util;
 
 pub use self::backstore::Backstore;
-pub use self::metadata::{SyncAll, MIN_MDA_SECTORS};
+pub use self::metadata::MIN_MDA_SECTORS;
 pub use self::setup::{find_all, get_metadata, is_stratis_device, setup_pool};
